requirements.txt
torch
tensorflow
pillow
onx
torchvision
onnxscript
optimum[exporters]
onnx_tf
flask
flask-restful